Make sample-signals asynchroneuos
This is a rather large change for (up until now) a rather
small feature: Instead of using own blocking calls, all
samples become values of Deferred.t - this leads to the simulation
to be a Deferred.t computation.

The benefit is typesafe asynchroneous control. In theory it should be
relatively easy now to start/stop/(maybe even rewind) a simulation.

Signed-off-by: Christoph Höger <christoph.hoeger@tu-berlin.de>